+++ pkg/security/module/cws.go
@@ -34,6 +34,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/security/seclog"
 	"github.com/DataDog/datadog-agent/pkg/security/telemetry"
 	"github.com/DataDog/datadog-agent/pkg/security/utils"
+	grpcutils "github.com/DataDog/datadog-agent/pkg/security/utils/grpc"
 )
 
 const (
@@ -59,7 +60,7 @@ type CWSConsumer struct {
 	rateLimiter    *events.RateLimiter
 	sendStatsChan  chan chan bool
 	eventSender    events.EventSender
-	grpcServer     *GRPCServer
+	grpcServer     *grpcutils.GRPCServer
 	ruleEngine     *rulesmodule.RuleEngine
 	selfTester     *selftests.SelfTester
 	selfTestCount  int
@@ -106,7 +107,7 @@ func NewCWSConsumer(evm *eventmonitor.EventMonitor, cfg *config.RuntimeSecurityC
 		apiServer:     apiServer,
 		rateLimiter:   events.NewRateLimiter(cfg, evm.StatsdClient),
 		sendStatsChan: make(chan chan bool, 1),
-		grpcServer:    NewGRPCServer(family, cfg.CmdSocketPath),
+		grpcServer:    grpcutils.NewGRPCServer(family, cfg.CmdSocketPath),
 		selfTester:    selfTester,
 		reloader:      NewReloader(),
 		crtelemetry:   crtelemetry,
@@ -144,7 +145,7 @@ func NewCWSConsumer(evm *eventmonitor.EventMonitor, cfg *config.RuntimeSecurityC
 	seclog.SetPatterns(cfg.LogPatterns...)
 	seclog.SetTags(cfg.LogTags...)
 
-	api.RegisterSecurityModuleServer(c.grpcServer.server, c.apiServer)
+	api.RegisterSecurityModuleServer(c.grpcServer.ServiceRegistrar(), c.apiServer)
 
 	// platform specific initialization
 	if err := c.init(evm, cfg, opts); err != nil {
